packages/core/icon-search/requirements.txt
torch
transformers
pillow
faiss-cpu
numpy
fastapi
uvicorn[standard]
pydantic

